+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to
 
 ### Changed
 
+- Do not throw if bigquery API returns
+  `The project ${config.projectId} has not enabled BigQuery.`
+
 - Rely on the individual steps to call service APIs instead of pre-calculating
   which service APIs are enabled in `getStepStartStates`. This should help
   drastically reduce the number of API calls to `serviceusage.googleapis.com`.